remove ports ; fix logstash filemounts.
The logstash filemounts needed to be renamed because there was an extra
logstash pipeline confif file that was causing every even to be written
to stdout, which eventually filled up disk. so MY pipeline file needs to
be mounted to "pipeline/logstash.conf" to fix this. The logstash YAML
settings file also needed to be "yml" instead of "yaml".

Also: logstash uses persisted queues so if the container is killed, it
will still have its queues availble.

Also: removed ports from containers that don't need them to be exposed.

Also: Added "LS_OPTS=--quiet", but i'm not sure it does anything.